update endpoint create collection

+++ pages/api/collection/save/[...path].js
@@ -1,10 +1,10 @@
 const matter = require('gray-matter')
 const yaml = require('yaml')
 const fs = require('fs')
-const collections = require('../../../lib/getCollection')
+const collections = require('../../../../lib/getCollection')
 
 const invalidCollection = (res, entryName) => {
-  res.statusCode = 400
+  res.statusCode = 404
   res.json({ message: `There is no ${entryName} collections!` })
   return
 }
@@ -12,6 +12,12 @@ const invalidCollection = (res, entryName) => {
 const success = res => {
   res.statusCode = 201
   res.json({ message: 'Successfully saved!' })
+  return
+}
+
+const notFound = res => {
+  res.statusCode = 404
+  res.json({ message: 'Not found!' })
   return
 }
 
@@ -23,20 +29,23 @@ const error = res => {
 
 export default (req, res) => {
   if (req.method === 'POST') {
-    const { entry, data } = req.body
-    const collection = collections.default().find(({ name }) => name === entry.name)
+    const { data } = req.body
+    const [collectionName, slug] = req.query.path
+    const collection = collections.default().find(({ name }) => name === collectionName)
     if (typeof collection === 'undefined') {
-      invalidCollection(res, entry.name)
+      invalidCollection(res, collectionName)
     }
 
     try {
       if (collection.hasOwnProperty('folder')) {
+        if (slug && !fs.existsSync(`${collection.folder}/${slug}.md`)) return notFound(res)
+
         const { body, ...meta } = data
-        fs.existsSync(entry.file) && fs.unlinkSync(entry.file)
+        fs.existsSync(`${collection.folder}/${slug}.md`) && fs.unlinkSync(`${collection.folder}/${slug}.md`)
         fs.writeFileSync(`${collection.folder}/${meta.slug}.md`, matter.stringify(body ?? '', meta))
       }
       if (collection.hasOwnProperty('files')) {
-        const collectionFile = collection.files.find(({ name }) => name === entry.slug)
+        const collectionFile = collection.files.find(({ name }) => name === slug)
         fs.writeFileSync(`${collectionFile.file}`, yaml.stringify(data))
       }
       success(res)
